pyproject.toml
[tool.poetry]
name = "Fabricius"
version = "0.3.0.dev1"
description = "Fabricius: The supportive templating engine for Python!"
license = "MIT"
authors = ["Predeactor <pro.julien.mauroy@gmail.com>"]
readme = "README.md"
repository = "https://github.com/Predeactor/Fabricius"
documentation = "https://fabricius.readthedocs.org"
keywords = ["cookiecutter", "template", "project", "scaffold", "pyscaffold"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",

    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",

    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",

    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: Implementation :: CPython",
    "Operating System :: OS Independent",

    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Utilities",

    "Typing :: Typed",
]

packages = [
    { include = "fabricius", from = "src" },
]

include = ["README.md", "LICENSE"]

[tool.poetry.dependencies]
python = "^3.10"
chevron = "^0.14.0"
rich = "^13.3.3"
typing-extensions = "^4.2.0"
inflection = "^0.5.1"
click = "^8.1.3"
jinja2 = "^3.1.2"
jinja2-time = "^0.2.0"
slugify = "^0.0.1"
platformdirs = "^3.5.1"
clickloader = "^0.1.3"
pydantic = "^2.2.0"
gitpython = "^3.1.32"
aspreno = "^1.0.0.dev2"
typer = "^0.9.0"

[tool.poetry.scripts]
fabricius = 'fabricius.cli:main'

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
Sphinx = "^5.0.2"
furo = "^2022.6.21"
sphinx-autobuild = "^2021.3.14"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
pre-commit = "^2.19.0"
mypy = "^1.3.0"
tox = "^4.4.5"
types-chevron = "^0.14.2.2"
esbonio = "^0.16.1"
ipykernel = "^6.22.0"
pytest = "^7.3.1"
poethepoet = "^0.20.0"
pytest-pretty = "^1.2.0"
pytest-html = "^3.2.0"
pytest-cov = "^4.1.0"
ruff = "^0.0.272"
types-pyyaml = "^6.0.12.11"

[tool.poetry.group.actions]
optional = true

[tool.poetry.group.actions.dependencies]
tox-gh-actions = "^3.0.0"

[tool.poe]
poetry_command = ""

[tool.poe.tasks]
_format_black = "black ."
_format_isort = "isort ."
_ruff = "ruff ./src/fabricius"
_mypy = "mypy -p fabricius --enable-incomplete-feature=Unpack"
format = ["_format_black", "_format_isort"]
test = "pytest . --html=tests/report/index.html --cov-config=.coveragerc --cov-context=test --cov-report=term-missing --cov-report=html --cov=fabricius"
lint = ["_ruff", "_mypy"]

[tool.poe.tasks.docs]
cmd = "make livehtml"
cwd = "./docs"

[tool.poe.tasks.docscov]
cmd = "make html -b coverage"
cwd = "./docs"

[tool.black]
line-length = 99

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
